+++ ulauncher/main.py
@@ -11,10 +11,10 @@ import gi
 
 import ulauncher.utils.xinit  # noqa: F401 - must import this before any GUI libraries are initialized
 from ulauncher import api_version, paths, version
-from ulauncher.app import UlauncherApp
 from ulauncher.cli import get_cli_args
 from ulauncher.gi import GLib, Gtk
-from ulauncher.ui import layer_shell
+from ulauncher.ui.app import UlauncherApp
+from ulauncher.ui.helpers import layer_shell
 from ulauncher.utils.environment import DESKTOP_ID, DESKTOP_NAME, DISTRO, IS_X11_COMPATIBLE, XDG_SESSION_TYPE
 from ulauncher.utils.logging_color_formatter import ColoredFormatter
 from ulauncher.utils.migrate import v5_to_v6
