+++ ulauncher/ui/helpers/system_theme.py
