+++ ulauncher/ui/preferences/views/extensions.py
@@ -9,11 +9,12 @@ from ulauncher.modes.extensions.extension_controller import (
     ExtensionController,
     ExtensionPreference,
 )
-from ulauncher.ui.windows.preferences import views
-from ulauncher.ui.windows.preferences.utils import ext_utils
-from ulauncher.ui.windows.preferences.utils.ext_handlers import ExtensionHandlers
-from ulauncher.ui.windows.preferences.utils.sidebar_layout import SidebarItem, SidebarLayout
-from ulauncher.ui.windows.preferences.views import (
+from ulauncher.ui import load_icon_surface
+from ulauncher.ui.preferences import views
+from ulauncher.ui.preferences.utils import ext_utils
+from ulauncher.ui.preferences.utils.ext_handlers import ExtensionHandlers
+from ulauncher.ui.preferences.utils.sidebar_layout import SidebarItem, SidebarLayout
+from ulauncher.ui.preferences.views import (
     BaseView,
     TextArea,
     get_window_for_widget,
@@ -21,7 +22,6 @@ from ulauncher.ui.windows.preferences.views import (
     styled,
 )
 from ulauncher.utils.launch_detached import open_detached
-from ulauncher.utils.load_icon_surface import load_icon_surface
 
 logger = logging.getLogger()
 REFRESH_INTERVAL = 250
